# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MyIFAPortal Financial Advisor Sales Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  timestamp: string;
}

export interface DashboardSummary {
  activePolicies: number;
  upcomingAppointments: number;
  totalDocuments: number;
  accountStatus: string;
  pendingFnaCount: number;
  recentActivity: ActivityItem[];
}

export interface FnaSubmission {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  dob: string;
  gender: string;
  smoker: string;
  maritalStatus: string;
  dependants: number;
  currentAge: number;
  grossMonthlyIncome: number;
  netMonthlyIncome: number;
  spouseIncome: number;
  employmentType: string;
  occupation: string;
  hasGroupBenefits: string;
  monthlyExpenses: number;
  homeLoans: number;
  vehicleFinance: number;
  personalLoans: number;
  creditCards: number;
  otherDebts: number;
  savings: number;
  investments: number;
  retirementFunds: number;
  propertyValue: number;
  existingLifeCover: number;
  existingDisabilityCover: number;
  existingDreadDiseaseCover: number;
  targetRetirementAge: number;
  monthlyRetirementIncome: number;
  currentRetirementSavings: number;
  monthlyInvestmentBudget: number;
  investmentGoal: string;
  riskProfile: string;
  investmentHorizon: number;
  priorities: string;
  /** @nullable */
  notes?: string | null;
  status: string;
  /** @nullable */
  advisorId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface FnaInput {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  dob: string;
  gender: string;
  smoker: string;
  maritalStatus: string;
  dependants: number;
  currentAge: number;
  grossMonthlyIncome: number;
  netMonthlyIncome: number;
  spouseIncome: number;
  employmentType: string;
  occupation: string;
  hasGroupBenefits: string;
  monthlyExpenses: number;
  homeLoans: number;
  vehicleFinance: number;
  personalLoans: number;
  creditCards: number;
  otherDebts: number;
  savings: number;
  investments: number;
  retirementFunds: number;
  propertyValue: number;
  existingLifeCover: number;
  existingDisabilityCover: number;
  existingDreadDiseaseCover: number;
  targetRetirementAge: number;
  monthlyRetirementIncome: number;
  currentRetirementSavings: number;
  monthlyInvestmentBudget: number;
  investmentGoal: string;
  riskProfile: string;
  investmentHorizon: number;
  priorities: string;
  notes?: string;
  advisorId?: number;
}

export interface FnaStatusUpdate {
  status: string;
  notes?: string;
}

export interface Policy {
  id: number;
  policyNumber: string;
  type: string;
  clientName: string;
  coverAmount: number;
  premium: number;
  status: string;
  startDate: string;
  /** @nullable */
  renewalDate?: string | null;
}

export interface PolicyInput {
  policyNumber: string;
  type: string;
  clientName: string;
  coverAmount: number;
  premium: number;
  status: string;
  startDate: string;
  renewalDate?: string;
}

export interface PolicyUpdate {
  policyNumber?: string;
  type?: string;
  clientName?: string;
  coverAmount?: number;
  premium?: number;
  status?: string;
  startDate?: string;
  renewalDate?: string;
}

export interface Appointment {
  id: number;
  clientName: string;
  clientEmail: string;
  advisorId: number;
  advisorName: string;
  date: string;
  time: string;
  type: string;
  status: string;
  /** @nullable */
  notes?: string | null;
}

export interface AppointmentInput {
  clientName: string;
  clientEmail: string;
  advisorId: number;
  date: string;
  time: string;
  type: string;
  notes?: string;
}

export interface AppointmentUpdate {
  clientName?: string;
  clientEmail?: string;
  advisorId?: number;
  date?: string;
  time?: string;
  type?: string;
  status?: string;
  notes?: string;
}

export interface Document {
  id: number;
  name: string;
  type: string;
  clientName: string;
  uploadedAt: string;
  size: string;
  status: string;
}

export interface DocumentInput {
  name: string;
  type: string;
  clientName: string;
  size: string;
}

export interface DocumentUpdate {
  status?: string;
  name?: string;
}

export interface Advisor {
  id: number;
  name: string;
  title: string;
  email: string;
  phone: string;
  initials: string;
  specializations: string[];
  clientCount: number;
}

/**
 * @nullable
 */
export type AdminSessionAdmin = {
  id?: number;
  name?: string;
  email?: string;
  role?: string;
} | null;

export interface AdminSession {
  /** @nullable */
  admin: AdminSessionAdmin;
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface AdminDashboard {
  totalAdvisors: number;
  totalClients: number;
  totalFnaSubmissions: number;
  pendingFna: number;
  totalPolicies: number;
  totalAppointments: number;
  totalDocuments: number;
  recentFna: FnaSubmission[];
  recentAppointments: Appointment[];
}

export interface AdminAdvisor {
  id: number;
  name: string;
  title: string;
  email: string;
  phone: string;
  initials: string;
  specializations: string[];
  clientCount: number;
  createdAt: string;
}

export interface AdminAdvisorInput {
  name: string;
  title: string;
  email: string;
  phone: string;
  initials: string;
  specializations: string[];
  password: string;
}

export interface AdminAdvisorUpdate {
  name?: string;
  title?: string;
  email?: string;
  phone?: string;
  initials?: string;
  specializations?: string[];
  clientCount?: number;
  password?: string;
}

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  role: string;
  createdAt: string;
}

export interface AdminUserInput {
  name: string;
  email: string;
  password: string;
  role: string;
}

export interface AdminUserUpdate {
  name?: string;
  email?: string;
  role?: string;
  password?: string;
}

export interface AdminFnaUpdate {
  status: string;
  /** @nullable */
  advisorId?: number | null;
  notes?: string;
}

export interface Client {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}