# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MyIFAPortal Financial Advisor Sales Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/advisors:
    get:
      operationId: adminListAdvisors
      tags: [admin-advisors]
      summary: List all advisors
      responses:
        "200":
          description: List of advisors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminAdvisor"
    post:
      operationId: adminCreateAdvisor
      tags: [admin-advisors]
      summary: Create a new advisor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAdvisorInput"
      responses:
        "201":
          description: Advisor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAdvisor"
  /admin/advisors/{id}:
    get:
      operationId: adminGetAdvisor
      tags: [admin-advisors]
      summary: Get advisor details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Advisor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAdvisor"
        "404":
          description: Not found
    patch:
      operationId: adminUpdateAdvisor
      tags: [admin-advisors]
      summary: Update advisor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAdvisorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAdvisor"
    delete:
      operationId: adminDeleteAdvisor
      tags: [admin-advisors]
      summary: Delete advisor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Admin Admins (super_admin only) ─────────────────────────────────────────
...
  /admin/fna/{id}:
    get:
      operationId: adminGetFna
      tags: [admin-fna]
      summary: Get FNA submission details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: FNA submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FnaSubmission"
        "404":
          description: Not found
    patch:
      operationId: adminUpdateFna
      tags: [admin-fna]
      summary: Update FNA (status, advisor assignment, notes)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminFnaUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FnaSubmission"
    delete:
      operationId: adminDeleteFna
      tags: [admin-fna]
      summary: Delete FNA submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Admin Appointments ───────────────────────────────────────────────────────
...
  /admin/policies:
    get:
      operationId: adminListPolicies
      tags: [admin-policies]
      summary: List all policies
      responses:
        "200":
          description: List of policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Policy"
    post:
      operationId: adminCreatePolicy
      tags: [admin-policies]
      summary: Create policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
  /admin/policies/{id}:
    patch:
      operationId: adminUpdatePolicy
      tags: [admin-policies]
      summary: Update policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
    delete:
      operationId: adminDeletePolicy
      tags: [admin-policies]
      summary: Delete policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Admin Documents ──────────────────────────────────────────────────────────
...
  /admin/clients/{id}:
    get:
      operationId: adminGetClient
      tags: [admin-clients]
      summary: Get client details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    delete:
      operationId: adminDeleteClient
      tags: [admin-clients]
      summary: Delete client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    FnaSubmission:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        dob:
          type: string
        gender:
          type: string
        smoker:
          type: string
        maritalStatus:
          type: string
        dependants:
          type: integer
        currentAge:
          type: integer
        grossMonthlyIncome:
          type: number
        netMonthlyIncome:
          type: number
        spouseIncome:
          type: number
        employmentType:
          type: string
        occupation:
          type: string
        hasGroupBenefits:
          type: string
        monthlyExpenses:
          type: number
        homeLoans:
          type: number
        vehicleFinance:
          type: number
        personalLoans:
          type: number
        creditCards:
          type: number
        otherDebts:
          type: number
        savings:
          type: number
        investments:
          type: number
        retirementFunds:
          type: number
        propertyValue:
          type: number
        existingLifeCover:
          type: number
        existingDisabilityCover:
          type: number
        existingDreadDiseaseCover:
          type: number
        targetRetirementAge:
          type: integer
        monthlyRetirementIncome:
          type: number
        currentRetirementSavings:
          type: number
        monthlyInvestmentBudget:
          type: number
        investmentGoal:
          type: string
        riskProfile:
          type: string
        investmentHorizon:
          type: integer
        priorities:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
        advisorId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - firstName
        - lastName
        - email
        - phone
        - dob
        - gender
        - smoker
        - maritalStatus
        - dependants
        - currentAge
        - grossMonthlyIncome
        - netMonthlyIncome
        - spouseIncome
        - employmentType
        - occupation
        - hasGroupBenefits
        - monthlyExpenses
        - homeLoans
        - vehicleFinance
        - personalLoans
        - creditCards
        - otherDebts
        - savings
        - investments
        - retirementFunds
        - propertyValue
        - existingLifeCover
        - existingDisabilityCover
        - existingDreadDiseaseCover
        - targetRetirementAge
        - monthlyRetirementIncome
        - currentRetirementSavings
        - monthlyInvestmentBudget
        - investmentGoal
        - riskProfile
        - investmentHorizon
        - priorities
        - status
        - createdAt
        - updatedAt
...
    Policy:
      type: object
      properties:
        id:
          type: integer
        policyNumber:
          type: string
        type:
          type: string
        clientName:
          type: string
        coverAmount:
          type: number
        premium:
          type: number
        status:
          type: string
        startDate:
          type: string
        renewalDate:
          type: ["string", "null"]
      required:
        - id
        - policyNumber
        - type
        - clientName
        - coverAmount
        - premium
        - status
        - startDate
    PolicyInput:
      type: object
      properties:
        policyNumber:
          type: string
        type:
          type: string
        clientName:
          type: string
        coverAmount:
          type: number
        premium:
          type: number
        status:
          type: string
        startDate:
          type: string
        renewalDate:
          type: string
      required:
        - policyNumber
        - type
        - clientName
        - coverAmount
        - premium
        - status
        - startDate
    PolicyUpdate:
      type: object
      properties:
        policyNumber:
          type: string
        type:
          type: string
        clientName:
          type: string
        coverAmount:
          type: number
        premium:
          type: number
        status:
          type: string
        startDate:
          type: string
        renewalDate:
          type: string
...
    AdminAdvisor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
        email:
          type: string
        phone:
          type: string
        initials:
          type: string
        specializations:
          type: array
          items:
            type: string
        clientCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - title
        - email
        - phone
        - initials
        - specializations
        - clientCount
        - createdAt
    AdminAdvisorInput:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        email:
          type: string
        phone:
          type: string
        initials:
          type: string
        specializations:
          type: array
          items:
            type: string
        password:
          type: string
      required:
        - name
        - title
        - email
        - phone
        - initials
        - specializations
        - password
    AdminAdvisorUpdate:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        email:
          type: string
        phone:
          type: string
        initials:
          type: string
        specializations:
          type: array
          items:
            type: string
        clientCount:
          type: integer
        password:
          type: string
...
    AdminFnaUpdate:
      type: object
      properties:
        status:
          type: string
        advisorId:
          type: ["integer", "null"]
        notes:
          type: string
      required:
        - status
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - createdAt