# lib/api-zod/src/generated/types/adminAdvisorUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MyIFAPortal Financial Advisor Sales Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminAdvisorUpdate {
  name?: string;
  title?: string;
  email?: string;
  phone?: string;
  initials?: string;
  specializations?: string[];
  clientCount?: number;
  password?: string;
}