# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MyIFAPortal Financial Advisor Sales Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Client {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}