# lib/api-zod/src/generated/types/policyInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MyIFAPortal Financial Advisor Sales Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface PolicyInput {
  policyNumber: string;
  type: string;
  clientName: string;
  coverAmount: number;
  premium: number;
  status: string;
  startDate: string;
  renewalDate?: string;
}